requirements/base.txt


# Bleeding edge Django
django==1.9.8

# Configuration
django-environ==0.4.0
whitenoise==3.2


# Forms
django-braces==1.9.0
django-crispy-forms==1.6.0
django-floppyforms==1.6.2

# Models
django-model-utils==2.5

# Images
Pillow==3.3.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.26.1


# Python-PostgreSQL Database Adapter
psycopg2==2.6.2

# Unicode slugification
unicode-slugify==0.1.3
django-autoslug==1.9.3

# Time zones support
pytz==2016.6.1

# Redis support
django-redis==4.4.4
redis>=2.10.0


celery==3.1.23






# Your custom requirements go here
MySQL-python==1.2.5
numpy==1.11.1
pandas==0.18.1
xlrd==1.0.0
django-chartit==0.2.6